backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database dependencies
psycopg2-binary==2.9.7
pgvector==0.2.3

# FREE AI/ML dependencies (replaces OpenAI)
transformers==4.35.2
torch>=2.6.0
sentence-transformers==2.2.2
google-generativeai==0.3.2

# Image processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.0
bcrypt==4.0.1
email-validator==2.1.0
schedule==1.2.0

# REMOVED (replaced with free alternatives):
# sendgrid==6.10.0  # Replaced with Gmail SMTP (built-in Python)
# boto3==1.29.7     # Replaced with local storage
# openai==1.3.5     # Replaced with Hugging Face transformers